release

+++ my-common/src/main/java/indi/uhyils/pojo/request/DateRangeRequest.java
@@ -11,27 +11,26 @@ public class DateRangeRequest extends DefaultRequest {
     /**
      * 开始时间
      */
-    private Integer start;
+    private Long start;
 
     /**
      * 结束时间
      */
-    private Integer end;
+    private Long end;
 
-
-    public Integer getStart() {
+    public Long getStart() {
         return start;
     }
 
-    public void setStart(Integer start) {
+    public void setStart(Long start) {
         this.start = start;
     }
 
-    public Integer getEnd() {
+    public Long getEnd() {
         return end;
     }
 
-    public void setEnd(Integer end) {
+    public void setEnd(Long end) {
         this.end = end;
     }
 }
